vivek/requirement.txt
pandas
numpy
scikit-learn
joblib
Pillow